Reorganise SnippetEditor
