+++ packages/search-metadata-previews/src/Shared.js
