+++ packages/search-metadata-previews/src/SnippetEditor/__mocks__/ReplacementVariableEditorStandalone.js
