--- packages/search-metadata-previews/src/SnippetEditor/index.js
@@ -1,13 +1,0 @@
-// Internal dependencies.
-import SnippetEditor from "./components/SnippetEditor";
-import SettingsSnippetEditor from "./components/SettingsSnippetEditor";
-import ReplaceVarEditor from "./components/ReplacementVariableEditor";
-
-export {
-	SnippetEditor,
-	ReplaceVarEditor,
-	SettingsSnippetEditor,
-};
-
-export * from "./constants";
-export * from "./components/Shared";
