+++ packages/search-metadata-previews/src/helpers/positionSuggestions.js
