+++ packages/search-metadata-previews/src/helpers/replaceText.js
